extract proxy into a plugin
