+++ wazo_phoned/controller.py
@@ -8,7 +8,6 @@ from xivo.token_renewer import TokenRenewer
 from wazo_auth_client import Client as AuthClient
 
 from .http_server import api, HTTPServer
-from .http import DirectoriesConfiguration
 
 logger = logging.getLogger(__name__)
 
@@ -18,7 +17,6 @@ class Controller:
         self.config = config
         self.http_server = HTTPServer(self.config['rest_api'])
         self.http_server.app.config['authorized_subnets'] = self.config['rest_api']['authorized_subnets']
-        DirectoriesConfiguration(config['dird'])
         self.token_renewer = TokenRenewer(self._new_auth_client(config))
         self.token_renewer.subscribe_to_token_change(self._on_token_change)
         plugin_helpers.load(
