+++ wazo_phoned/plugins/api/resource.py
