+++ wazo_phoned/plugins/proxy/resource.py
@@ -12,73 +12,19 @@ from flask import (
 from requests.exceptions import RequestException
 from time import time
 
-from xivo.mallow import fields
-from xivo.mallow_helpers import Schema
+from wazo_phoned.auth_remote_addr import AuthResource
 
-from .auth_remote_addr import AuthResource
-from .http_server import api
+from .exceptions import WazoAuthConnectionError, WazoDirdConnectionError
+from .schema import UserUUIDSchema, LookupSchema, LookupGigasetSchema
 
-logger = logging.getLogger(__name__)
 
 DIRD_API_VERSION = '0.1'
 
-
-class UserUUIDSchema(Schema):
-    xivo_user_uuid = fields.String(required=True)
-
-
-class LookupSchema(UserUUIDSchema):
-    term = fields.String(required=True)
-    limit = fields.Integer(missing=None)
-    offset = fields.Integer(missing=None)
-
-
-class LookupGigasetSchema(Schema):
-    set_first = fields.String(attribute='term', missing='')
-    count = fields.Integer(attribute='limit', missing=None)
-    first = fields.Integer(attribute='first', missing=1)
-
-
-class WazoAuthConnectionError(RequestException):
-
-    code = 503
-
-    def __str__(self):
-        return 'Connection to XiVO Auth failed'
-
-
-class WazoDirdConnectionError(RequestException):
-
-    code = 503
-
-    def __str__(self):
-        return 'Connection to XiVO Dird failed'
+logger = logging.getLogger(__name__)
 
 
 def _error(code, msg):
     return {'reason': [msg], 'timestamp': [time()], 'status_code': code}, code
-
-
-class DirectoriesConfiguration:
-
-    menu_url = '/directories/menu/<profile>/<vendor>'
-    input_url = '/directories/input/<profile>/<vendor>'
-    lookup_url = '/directories/lookup/<profile>/<vendor>'
-    lookup_gigaset_url = '/directories/lookup/<profile>/gigaset/<xivo_user_uuid>'
-
-    def __init__(self, dird_config):
-        dird_host = dird_config['host']
-        dird_port = dird_config['port']
-        dird_verify_certificate = dird_config.get('verify_certificate', True)
-
-        Menu.configure(dird_host, dird_port, dird_verify_certificate)
-        Input.configure(dird_host, dird_port, dird_verify_certificate)
-        Lookup.configure(dird_host, dird_port, dird_verify_certificate)
-        LookupGigaset.configure(dird_host, dird_port, dird_verify_certificate)
-        api.add_resource(Menu, self.menu_url)
-        api.add_resource(Input, self.input_url)
-        api.add_resource(Lookup, self.lookup_url)
-        api.add_resource(LookupGigaset, self.lookup_gigaset_url)
 
 
 class Menu(AuthResource):
